Add property-based tests for signers-voting contract functions
- reward-cycle-to-burn-height
- burn-height-to-reward-cycle
- is-in-prepare-phase

This effort, initially part of https://github.com/stacks-network/stacks-core/pull/4286,
is now in a separate commit and PR as it's autonomous with regards to PoX-4
property and fuzz testing. Additional property-based tests for the remaining
functions of the signers-voting contract are warranted and will be addressed
in separate commits/PRs.